Added abstract class
